engine/veriframe_core/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "veriframe-core"
version = "0.1.0"
description = "Local VeriFrame Python engine with FastAPI sidecar and contracts."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "torch",
  "torchvision",
  "opencv-python",
  "pillow",
  "numpy",
  "pydantic>=2",
  "fastapi",
  "uvicorn",
  "sqlalchemy",
  "httpx"
]

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-cov",
  "ruff",
  "mypy",
  "pyright"
]

[project.scripts]
veriframe = "veriframe_core.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["veriframe_core*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "S", "C4", "SIM", "RUF"]
ignore = ["S101"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
